requirements.txt
agno
python-dotenv
pydantic
openai
pandas